backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "researchpilot-backend"
version = "0.1.0"
description = "FastAPI backend for the ResearchPilot multi-agent research copilot."
requires-python = ">=3.11"
dependencies = [
  "dspy-ai>=2.5.0",
  "fastapi>=0.115.0",
  "httpx>=0.28.0",
  "pydantic>=2.9.0",
  "pydantic-settings>=2.5.0",
  "qdrant-client>=1.12.0",
  "uvicorn[standard]>=0.32.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

